requirements.txt
fastapi
uvicorn
transformers==4.35.2
transformers[sentencepiece]
torch
sacremoses
pydantic_settings
pytest